Fix overflow bug in fixed numbers

+++ core/math/fixed.h
@@ -5,5 +5,5 @@
 static const int fixed_k = 1000;
 typedef int64_t fixed;
 
-#define FIXED_ENCODE(x) ((x) * fixed_k)
+#define FIXED_ENCODE(x) (((int64_t)(x)) * fixed_k)
 #define FIXED_DECODE(x) ((x) / fixed_k)
